Fix implemented such that InterpolateData can be passed into PipelineData correctly.
